chore: checkpoint vanilla parity implementation

+++ crates/pumpkin/src/block/blocks/coral/mod.rs
@@ -49,7 +49,7 @@ async fn try_schedule_die_tick(block: &Block, world: &Arc<World>, pos: &BlockPos
     world.schedule_block_tick(
         block,
         *pos,
-        tick_delay as u8,
+        tick_delay as u32,
         pumpkin_world::tick::TickPriority::Normal,
     );
 }
